Fixing ldshell linking issue
